# docs/swagger.yaml
definitions:
  httputil.Error:
    properties:
      code:
        example: 404
        type: integer
      data: {}
      message:
        example: not found
        type: string
    type: object
  httputil.Message:
    properties:
      code:
        example: 200
        type: integer
      data: {}
      message:
        example: a server message
        type: string
    type: object
  models.AccessToken:
    properties:
      access_token:
        type: string
      expires_in:
        type: integer
      token_type:
        type: string
    type: object
  models.Role:
    enum:
    - admin
    - user
    type: string
    x-enum-varnames:
    - RoleAdmin
    - RoleUser
  models.User:
    properties:
      email:
        description: Name             string    `json:"name"                          gorm:"type:varchar(255);not
          null"`
        type: string
      id:
        type: string
      reset_password_token:
        type: string
      role:
        $ref: '#/definitions/models.Role'
      verification_code:
        type: string
    type: object
  models.UserPasswordConfirmation:
    properties:
      password:
        minLength: 8
        type: string
      password_confirm:
        type: string
    required:
    - password
    - password_confirm
    type: object
  models.UserSignIn:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  models.UserSignUp:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        minLength: 8
        type: string
      password_confirm:
        type: string
    required:
    - email
    - name
    - password
    - password_confirm
    type: object
info:
  contact: {}
paths:
  /auth/password/change:
    patch:
      consumes:
      - application/json
      description: Change the user password
      parameters:
      - description: New password
        in: body
        name: password
        required: true
        schema:
          $ref: '#/definitions/models.UserPasswordConfirmation'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httputil.Message'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputil.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httputil.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.Error'
        "412":
          description: Precondition Failed
          schema:
            $ref: '#/definitions/httputil.Error'
      summary: Change the user password
      tags:
      - auth
  /auth/password/forgot/{email}:
    get:
      consumes:
      - application/json
      description: Send a reset token by email
      parameters:
      - description: User email
        format: email
        in: path
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httputil.Message'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputil.Error'
      summary: Send a reset token by email
      tags:
      - auth
  /auth/password/reset/{reset_token}:
    patch:
      consumes:
      - application/json
      description: Reset the user password
      parameters:
      - description: New password
        in: body
        name: password
        required: true
        schema:
          $ref: '#/definitions/models.UserPasswordConfirmation'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httputil.Message'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputil.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httputil.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.Error'
        "412":
          description: Precondition Failed
          schema:
            $ref: '#/definitions/httputil.Error'
      summary: Reset the user password
      tags:
      - auth
  /auth/refresh:
    get:
      consumes:
      - application/json
      description: Refresh the access token
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AccessToken'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httputil.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.Error'
      summary: Refresh the access token
      tags:
      - auth
  /auth/signin:
    post:
      consumes:
      - application/json
      description: Sign in a user
      parameters:
      - description: User credential
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/models.UserSignIn'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AccessToken'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.Error'
      summary: Sign in a user
      tags:
      - auth
  /auth/signout:
    get:
      consumes:
      - application/json
      description: Sign out current user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httputil.Message'
      summary: Sign out current user
      tags:
      - auth
  /auth/signup:
    post:
      consumes:
      - application/json
      description: Sign up a new user
      parameters:
      - description: User sign up
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.UserSignUp'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.Error'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/httputil.Error'
        "412":
          description: Precondition Failed
          schema:
            $ref: '#/definitions/httputil.Error'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/httputil.Error'
      summary: Sign up a new user
      tags:
      - auth
  /auth/verify/{verification_code}:
    get:
      consumes:
      - application/json
      description: Verify email address from verification code sent by email
      parameters:
      - description: verification code sent by email
        in: path
        name: verification_code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httputil.Message'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.Error'
      summary: Verify email address
      tags:
      - auth
  /auth/welcome/{email}:
    get:
      consumes:
      - application/json
      description: This re-sends the welcome email to the user if the user is not
        verified
      parameters:
      - description: User email
        format: email
        in: path
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httputil.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.Error'
      summary: Send welcome email
      tags:
      - auth
  /users/me:
    get:
      consumes:
      - application/json
      description: Get the current user
      produces:
      - application/json
      responses: {}
      summary: Get the current user
      tags:
      - user
swagger: "2.0"